[FEATURE] Add support for "img" and "callback" services
A "img" service creates an <img> tag instead of an
iframe or script. Useful for external images
like static google maps

The "callback" service does not create any tag but
executes a JS callback function to handle the embeding
itself. This is supposed for more complex embeds like
a custom google maps script